Profile Page Skeleton + Concept
Most of the work for it is trying to make things scale dynamically, and a lot of it lacks any meaningful features so far.

Saving doing work on displaying quests until I see that system myself and can implement it. However things like the skills list or list of classes later im already setting up ways to dynamically add as many elements are are found and pulled out of our databases