Profile Page / Quests / Courses integration
Extremely bare bones for now (and built from a lot of things in the develop branch!

My work consists primarily of things regarding profile / profileS right now, and the required edits to urls / views to make that function.

Still more to, I want to add the skill types next to the models and such? Perhaps?